major renaming and new  3D demos for Matlab
