+++ Core/regularisers_CPU/FGP_TV_core.c
@@ -255,7 +255,7 @@ float Grad_func3D(float *P1, float *P2, float *P3, float *D, float *R1, float *R
 {
     float val1, val2, val3, multip;
     int i,j,k, index;
-    multip = (1.0f/(8.0f*lambda));
+    multip = (1.0f/(26.0f*lambda));
 #pragma omp parallel for shared(P1,P2,P3,D,R1,R2,R3,multip) private(index,i,j,k,val1,val2,val3)
     for(i=0; i<dimX; i++) {
         for(j=0; j<dimY; j++) {
