+++ Core/regularisers_GPU/TV_ROF_GPU_core.h
