+++ Wrappers/Matlab/mex_compile/compileCPU_mex.m
@@ -1,10 +1,10 @@
 % execute this mex file in Matlab once
-copyfile ../../../Core/regularizers_CPU/ regularizers_CPU/
-copyfile ../../../Core/CCPiDefines.h regularizers_CPU/
+copyfile ../../../Core/regularisers_CPU/ regularisers_CPU/
+copyfile ../../../Core/CCPiDefines.h regularisers_CPU/
 
-cd regularizers_CPU/
+cd regularisers_CPU/
 
-fprintf('%s \n', 'Compiling CPU regularizers...');
+fprintf('%s \n', 'Compiling CPU regularisers...');
 mex ROF_TV.c ROF_TV_core.c utils.c CFLAGS="\$CFLAGS -fopenmp -Wall -std=c99" LDFLAGS="\$LDFLAGS -fopenmp"
 movefile ROF_TV.mex* ../installed/
