--- Wrappers/Python/fista-recipe/meta.yaml
@@ -1,29 +1,0 @@
-package:
-  name: ccpi-fista
-  version: {{ environ['CIL_VERSION'] }}
-
-
-build:
-  preserve_egg_dir: False
-  script_env:
-    - CIL_VERSION   
-#  number: 0
-  
-requirements:
-  build:
-    - python
-    - numpy
-    - setuptools
- 
-  run:
-    - python
-    - numpy
-    #- astra-toolbox
-    - ccpi-regularizer
-
-
-	
-about:
-  home: http://www.ccpi.ac.uk
-  license:  Apache v.2.0 license
-  summary: 'CCPi Core Imaging Library (Viewer)'
